Validation Enchance
As a result, when a user requests this JSP for the first time, no validation message is displayed, and empty strings are displayed in the three text boxes of the form. However, if the user doesn't enter a valid string for each of the three text boxes, this JSP displays a validation message in italics before the form, and it displays any values that the user has already entered within the text boxes.

page 220 of Murach's JSP